requirements.txt
moviepy==1.0.3
gtts==2.3.2
requests==2.31.0
pandas==2.0.3
openai==1.12.0
python-dotenv==1.0.0

# YouTube API dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0